meta: add makefile, test runner, skeleton MessageBuffer
Add the very beginnings of MessageBuffer, which will
manage a buffer of log messages. Along the way, add
the support files that we need to build and test
MessageBuffer.

Note that the core functionality of MessageBuffer will be
added in later CLs. (The rest of the functionality is too
big to fit in this CL.)

Bug: 31653003
Test: ./runtests.sh (on bullhead)

Change-Id: I6c8a1bf453d02438988c955d70196e67f8173d1d