add command processor
Add CommandProcessor, the component which will parse incoming
protocol commands, and take the appropriate action. This initial
implementation of CommandProcessor handles on
protocol::kWriteAsciiMessage commands.

Add unit tests for CommandProcessor as well. Note that these
unit tests can't check that the messages were actually stored,
because CommandProcessor doesn't yet provide the ability to read
the stored data.

As such, we'll add tests that verify that our logs were stored,
when we add dump functionality to CommandProcessor.

Bug: 31861976
Test: ./runtest.sh (on bullhead)
Change-Id: I2f1eabbf4686f74d834e6837ec0ad5d24145f007