add ByteBuffer
ByteBuffer provides a simple way of concatenating data, such
as we might want to do when generating a protocol packet.

ByteBuffer is implemented as a class template, rather than a
simple class, because we want to be able to allocate a ByteBuffer
on the stack. For stack allocation, the object size must be
known at compile-time. Hence, ByteBuffer is a class template,
parameterized on the buffer size.

Bug: 31861967
Test: ./runtests.sh (on bullhead)
Change-Id: I1fe89a9cdcf34cf380365b3aad2537edca87a424