MessageBuffer: use CopyFromBufferOrDie()
Update ConsumeNextMessage() to make use of
local_utils::CopyFromBufferOrDie(). This (hopefully)
accomplishes two things:
- makes the code a bit more obvious, and
- makes the code more consistent with CommandProcessor

Along the way: fix a bug in the way that we check
that the MessageBuffer contains enough bytes for
a LengthHeader.

(Previously, we were checking that the LengthHeader
fit within the memory bounds of |data_|. Really,
though, we should check that the LengthHeader fits
within the populated area of |data_|.)

Bug: 31971382
Test: ./runtests.sh (on bullhead)
Change-Id: Ie3a4c38fefb69a3212142a53e79fdc720af6c601