CommandProcessor: rename parameters in tests
The BuildAsciiMessageCommandWithSizeAdjustments() and
SendAsciiMessageWithSizeAdjustments() methods of the
CommandProcessorTest fixture take parameters which are
used to tweak the values of various header fields.

Unfortunately, however, the mapping between the parameter
names and the header fields is unclear. Revise the
parameter names to make the mapping obvious. In particular,
the new names follow the pattern <struct-name>_<struct-field>.

Bug: 32255492
Test: ./run_tests.sh (on angler)
Change-Id: I18a0938a946a36e72271f8e00afa8af4800916ce