CommandProcessor: refactor fake Write() in unit tests
A number of test cases share a common implementation
of a fake Os::Write(). Move this implementation to
the CommandProcessorTest fixture, to reduce code
duplication.

However, to avoid too much reliance on implicit
behavior from the test fixture, the fixture does
not set any EXPECTations for the Os::Write() method.
Instead, the fixture simply sets a default action.

While there: enhance the ...IdempotentEvenWithWriteFailure
test. The new version of this test checks not only that
Write() is called on the second dump request, but that
some data is provided to that Write() call.

Bug: 32277627
Test: ./runtests.sh (on angler)
Change-Id: Ida8d48ae2b7d02d38834ef75141ed1c36209d54d