CommandProcessor: ProcessInput(): improve opcode handling
At present, unrecognized opcodes are silently ignored.
That will make it hard to debug issues with buggy or
corrupted clients.

Add logging for the unrecognized opcode case, and
leave a note to add a stats counter as well.

Note that we use DEBUG level logging, because bad
opcodes are unlikely to cause an externally observable
problems in normal operation. (In normal operation, no
one looks at logs.)

Bug: 32242567
Test: ./runtests.sh (on angler)
Change-Id: Idd447cc5dbf44c8853b9227afae4233ac50e1cf0