protocol: add setters for AsciiMessage
- Add setters for the fields of protocol::AsciiMessage,
  so that we can use chaining to initialize an
  AsciiMessage.
- Update the command processor unittests, to make
  use of this ability.

Note that this CL introduces a behavioral change.
Previously, the EXPECTations in
BuildAsciiMessageCommandWithAdjustments were based
on the unadjusted size of the ASCII message.

After this CL, the relevant EXPECTations are based
on the adjusted size. This is intentional, as our
goal is to verify that the AsciiMessage we construct
has the final (adjusted) size that the test case
would intend.

Bug: 32327379
Test: ./runtests.sh (on angler)
Change-Id: Ic568ae663de713f31d745635d278089e25359693