add host unit tests
Add the ability to compile and run unit tests for the host.
This enables development even when we don't have an Android
device at hand.

The build rules here are a bit asymmetric, since we link the
device unit tests against libwifilogd, but don't do the
same for host unit tests.

The rationale for this asymmetry is that, in the host case,
we don't need a library, since we're not going to build
the actual daemon. (For the host, we're only going to build
the tests.)

Bug: 32650144
Test: ./runtests.sh (with angler)
Change-Id: I1d458a6af66b50d1dd9c3ef73cb468db0cb82505